Move Error to scylla-cql-core
In general we don't want this type in core, but unfortunately we have
`impl From<response::error::Error> for RequestAttemptError` in scylla
crate.
Moving it required also moving a few utils from types module. Note that
`read_short_length` became public because of this.
